% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auoplot.R
\name{auto_plot}
\alias{auto_plot}
\title{Auto Plot for a single vector}
\usage{
auto_plot(x, colname = NULL)
}
\arguments{
\item{x}{A numeric vector or a factor/character vector (one column).}

\item{colname}{Optional character: column name (used in titles).}
}
\value{
Invisible NULL. Function draws plots to the active graphics device.
}
\description{
Draws a small set of exploratory plots for a numeric or categorical vector.
}
\examples{
auto_plot(iris)
}
