% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-list.R
\name{autoplot.list}
\alias{autoplot.list}
\title{Automatic ggplot for a x, y, z list.}
\usage{
\method{autoplot}{list}(object, data = NULL, mapping = aes(), ...)
}
\arguments{
\item{object}{list with components x, y and z}

\item{data}{not used by this method}

\item{mapping}{a call to aes() specifying additional mappings between variables and plot aesthetics; by default, positions in x and y are mapped to the x and y coordinates and fill is mapped to the values contained in the z component in the list}

\item{...}{not used by this method}
}
\description{
Similar to \code{\link[graphics]{image}}
}
\examples{
x <- seq(0, 1, by=0.1)
y <- seq(0, 2, by=0.1)
z <- outer(x, y, "+") 
d <- list(x=x, y=y, z=z)
image(d)
autoplot(d)
autoplot(d, mapping=aes(alpha=x))
}

