% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_autoplot.R
\name{autoplot_pca}
\alias{autoplot.PCA}
\alias{autoplot.pca}
\alias{autoplot.pcaRes}
\alias{autoplot.prcomp}
\alias{autoplot.rda}
\alias{autoplot_pca}
\title{Automatic ggplot for a Principal Component Analysis}
\usage{
\method{autoplot}{prcomp}(object, mapping = aes(), data = NULL,
  dimensions = c(1, 2), which = "rows", scaling = which,
  n.max.labels = 100, ...)

\method{autoplot}{PCA}(object, mapping = aes(), data = NULL,
  dimensions = c(1, 2), which = "rows", scaling = which,
  n.max.labels = 100, ...)

\method{autoplot}{rda}(object, mapping = aes(), data = NULL,
  dimensions = c(1, 2), which = "rows", scaling = which,
  n.max.labels = 100, ...)

\method{autoplot}{pca}(object, mapping = aes(), data = NULL,
  dimensions = c(1, 2), which = "rows", scaling = which,
  n.max.labels = 100, ...)

\method{autoplot}{pcaRes}(object, mapping = aes(), data = NULL,
  dimensions = c(1, 2), which = "rows", scaling = which,
  n.max.labels = 100, ...)
}
\arguments{
\item{object}{an object returned by a function performing Principal Component Analysis.}

\item{mapping}{a call to aes() specifying additional mappings between variables and plot aesthetics. By default, positions \code{x} and \code{y} are mapped to the scores on the principal components.}

\item{data}{the original dataset, to be concatenated with the output when extracting row scores. When \code{NULL} (the default) data will be extracted from the PCA object when it contains it (i.e. for all cases but \code{\link{prcomp}}).}

\item{dimensions}{vector giving the numbers of the principal components to extract. Typically two are extracted to create a plot.}

\item{which}{the plot to produce: either plot "rows", "lines", "observations", "objects", "individuals", "sites" (which are all treated as synonyms), or plot "columns", "variables", "descriptors", "species" (which are, again, synonyms), or produce a "biplot". All can be abbreviated. By default, observations are plotted.}

\item{scaling}{scaling for the scores. Can be
\describe{
  \item{"none" (or 0)}{for raw scores,}
  \item{"rows" (or 1, or a synonym of "rows")}{to scale row scores by the eigenvalues,}
  \item{"columns" (or 2, or a synonym of "columns")}{to scale column scores by the eigenvalues,}
  \item{"both" (or 3)}{to scale both row and column scores.}
}
By default, scaling is adapted to the type of scores extracted (scaling 1 for row scores, scaling 2 for column scores, and scaling 3 when scores are extracted for a biplot).}

\item{n.max.labels}{maximum number of observation labels to plot. Let \code{n} be the number of data rows. For \code{0 < n < n.max.labels/2}, labels are plotted and shifted to avoid over-plotting. For \code{n.max.labels/2 < n < n.max.labels}, labels are plotted but not disambiguated. For \code{n.max.labels < n} labels are not plotted.}

\item{...}{passed to the various geoms; can be used to \emph{set} further aesthetics.}
}
\value{
A ggplot2 object defining the plot.
}
\description{
Extract information from an object returned by a function performing Principal Component Analysis and produce a plot of observations, of variables, or a biplot.
}
\details{
The object is passed to the appropriate \code{augment} method, defined in \code{\link{pca_tidiers}}, which extracts the scores, possibly the original data, and other relevant information from the PCA object. The resulting \code{data.frame} is plotted with:
\describe{
  \item{\code{\link[ggplot2]{geom_point}}:}{for observations points,}
  \item{\code{\link[ggplot2]{geom_text}} or \code{\link[ggrepel]{geom_text_repel}}:}{for observations labels,}
  \item{\code{\link[ggplot2]{geom_segment}}:}{for variables vectors.}
}
}
\examples{
pca <- prcomp(USArrests, scale=TRUE)
autoplot(pca)
autoplot(pca, which="variables")
autoplot(pca, which="biplot") # defaults to scaling=3
autoplot(pca, which="biplot", scaling=1)
autoplot(pca, which="biplot", scaling=2, n.max.labels=0)

# add further aesthetic mappings
names(augment(pca, data=USArrests))
autoplot(pca, data=USArrests, mapping=aes(alpha=.cos2))
autoplot(pca, data=USArrests, mapping=aes(alpha=.cos2, size=.contrib))
# including from the original data
autoplot(pca, data=USArrests, mapping=aes(alpha=.cos2, color=Murder))

# aesthetics can also be set
autoplot(pca, mapping=aes(alpha=.cos2), color="darkblue", size=3)

if (require("FactoMineR")) {
  pca <- FactoMineR::PCA(USArrests, graph=FALSE)
  autoplot(pca)
  autoplot(pca, which="variables")
  # with FactoMineR, the data is present by default and can be mapped
  names(augment(pca))
  autoplot(pca, mapping=aes(alpha=.cos2, color=Murder))
}

if (require("vegan")) {
  pca <- vegan::rda(USArrests, scale=TRUE)
  plot(pca)
  autoplot(pca, which="biplot", scaling=2)
  autoplot(pca)
}

if (require("ade4")) {
  pca <- ade4::dudi.pca(USArrests, nf=4, scannf=FALSE)
  biplot(pca)
  autoplot(pca, which="biplot", scaling=1)
}

if (require("pcaMethods")) {
  pca <- pcaMethods::pca(USArrests, scale="uv", nPcs=4)
  biplot(pca)
  autoplot(pca, which="biplot")
}

}
\seealso{
Functions to perform PCA: \code{\link[stats]{prcomp}} in package \code{stats}, \code{\link[FactoMineR]{PCA}} in package \code{factoMineR}, \code{\link[vegan]{rda}} in package \code{vegan}, \code{\link[ade4]{dudi.pca}} in package \code{ade4}, \code{\link[pcaMethods]{pca}} in package \code{pcaMethods} (on bioconductor).

Other PCA-related functions: \code{\link{eigenvalues}},
  \code{\link{pca_tidiers}}, \code{\link{scores}}
}

