% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fortify-ca.R
\name{fortify_ca}
\alias{fortify.CA}
\alias{fortify.correspondence}
\alias{fortify_ca}
\title{Extract information from a Correspondence Analysis

Extract information from a Correspondence Analysis such as the scores on the principal components and some fit statistics into a data.frame.}
\usage{
\method{fortify}{correspondence}(model, data = NULL, PC = c(1, 2), ...)

\method{fortify}{CA}(model, data = NULL, PC = c(1, 2), ...)
}
\arguments{
\item{model}{an object resulting from a CA}

\item{data}{the original dataset, use to compute the contributions of rows and columns to the principal components. When \code{NULL}, the default, data will be extracted from the CA object when it contains it}

\item{PC}{the principal components to extract; two is usual for plotting}

\item{...}{pass-through argument}
}
\value{
A data.frame containing
  \item{.id}{the identifier of the rows (row name) or columns (column name)}
  \item{.PC#}{the scores of rows or columns on the extracted principal components}
  \item{.cos2}{the squared cosine summed over all extracted PCs}
  \item{.contrib}{the contribution to the selected PCs}
  \item{.kind}{the nature of the data extracted : rows, columns and possibly their status (active or supplementary)}
}
\description{

}
\author{
Jean-Olivier Irisson \email{irisson@normalesup.org}
}
\seealso{
\code{\link{autoplot_ca}} to produce plots based on the output of \code{fortify}.
}

