% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_tidiers.R
\name{pca_tidiers}
\alias{augment.PCA}
\alias{augment.pca}
\alias{augment.pcaRes}
\alias{augment.prcomp}
\alias{augment.rda}
\alias{pca_tidiers}
\alias{tidy.PCA}
\alias{tidy.pca}
\alias{tidy.pcaRes}
\alias{tidy.prcomp}
\alias{tidy.rda}
\title{Tidying methods for a Principal Component Analysis}
\usage{
\method{tidy}{prcomp}(x, ...)

\method{tidy}{PCA}(x, ...)

\method{tidy}{rda}(x, ...)

\method{tidy}{pca}(x, ...)

\method{tidy}{pcaRes}(x, ...)

\method{augment}{prcomp}(x, data = NULL, dimensions = c(1, 2),
  type = "row", scaling = type, ...)

\method{augment}{PCA}(x, data = NULL, dimensions = c(1, 2), type = "row",
  scaling = type, ...)

\method{augment}{pca}(x, data = NULL, dimensions = c(1, 2), type = "row",
  scaling = type, ...)

\method{augment}{pcaRes}(x, data = NULL, dimensions = c(1, 2),
  type = "row", scaling = type, ...)

\method{augment}{rda}(x, data = NULL, dimensions = c(1, 2), type = "row",
  scaling = type, ...)
}
\arguments{
\item{x}{an object returned by a function performing Principal Component Analysis.}

\item{...}{pass-through argument}

\item{data}{the original dataset, to be concatenated with the output when extracting row scores. When \code{NULL} (the default) data will be extracted from the PCA object when it contains it (i.e. for all cases but \code{\link{prcomp}})}

\item{dimensions}{vector giving the numbers of the principal components to extract. Typically two are extracted to create a plot.}

\item{type}{the scores (i.e. coordinates in the new space) to extract: either "rows", "lines", "observations", "objects", "individuals", "sites" (which are all treated as synonyms) or "columns", "variables", "descriptors", "species" (which are, again, synonyms). All can be abbreviated. By default, scores of rows are returned.}

\item{scaling}{scaling for the scores. Can be
\describe{
  \item{"none" (or 0)}{for raw scores,}
  \item{"rows" (or 1, or a synonym of "rows")}{to scale row scores by the eigenvalues,}
  \item{"column" (or 2, or a synonym of "column")}{to scale column scores by the eigenvalues,}
  \item{"both" (or 3)}{to scale both row and column scores.}
}
By default, scaling is adapted to the type of scores extracted (scaling 1 for row scores and scaling 2 for column scores).}
}
\value{
A data.frame containing the original data, when it is supplied or can be extracted from the object, and the additional columns
\describe{
  \item{.label}{the identifier of the row or column.}
  \item{.PC#}{the scores (coordinates) of the objects on the extracted principal components.}
  \item{.cos2}{the squared cosine summed over all extracted PCs, which quantifies the quality of the representation of the object on the extracted PCs.}
  \item{.contrib}{the contribution of each object to the selected PCs}
}
}
\description{
Extract information from a Principal Component Analysis such as the scores on the principal components and some fit statistics.
}

