% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{plot_bar}
\alias{plot_bar}
\title{Generate a bar plot with error bars, jitter points, and Tukey's letters}
\usage{
plot_bar(df, var, group, xlab, ylab)
}
\arguments{
\item{df}{a data frame}

\item{var}{a target variable to plot}

\item{group}{a grouping variable like Group or Treatment}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
a ggplot object
}
\description{
Generate a bar plot with error bars, jitter points, and Tukey's letters
}
\examples{
set.seed(123)
df <- data.frame(var = c(rnorm(9, 10, 4), rnorm(9, 20, 6), rnorm(9, 30, 9)),
                group = rep(c("Ctrl", "Pla", "DoSM"), each = 9))
plot_bar(df, "var", "group", "Treatment", "Cured number of patients")
}
