% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_bygroup.R
\name{scatter_bygroup}
\alias{scatter_bygroup}
\title{Scatter Plot of Two Continuous Variables Colored by Categorical Groups}
\usage{
scatter_bygroup(data, x, y, group, lm = TRUE)
}
\arguments{
\item{data}{Data frame containing the data.}

\item{x}{String of the first continuous variable name from the data frame to pass to
\code{tidyverse::ggplot()}. This is the x variable of the scatter plot.}

\item{y}{String of the second continuous variable name from the data frame to pass to
\code{tidyverse::ggplot()}. This is the y variable of the scatter plot.}

\item{group}{Character of name of categorical variable that points are be colored by}

\item{lm}{If \code{TRUE} (the default), uses the \code{geom_smooth()} function
to fit a linear regression line to the scatter plot. If \code{FALSE},
there is no linear regression line.}
}
\value{
A plot
}
\description{
Plot a scatter plot of two continuous variables and colored the points
by a categorical variable.
}
\examples{
library(datateachr)
scatter_bygroup(cancer_sample, "radius_mean", "concave_points_mean", "diagnosis")
scatter_bygroup(cancer_sample, "radius_mean", "concave_points_mean", "diagnosis",FALSE)
}
