% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot.list}
\alias{autoplot.list}
\title{Automatic ggplot for a x, y, z list.}
\usage{
\method{autoplot}{list}(object, mapping = aes(), ...)
}
\arguments{
\item{model}{list with components x, y and z}

\item{data}{not used by this method}

\item{...}{not used by this method}
}
\description{
Similar to \code{\link[graphics]{image}}
}
\examples{
x <- seq(0, 1, by=0.1)
y <- seq(0, 2, by=0.1)
z <- outer(x, y, "+")
d <- list(x=x, y=y, z=z)
image(d)
autoplot(d)
autoplot(d, mapping=aes(alpha=x))
}

