% Generated by roxygen2 (4.0.2): do not edit by hand
\name{autoplot_pca}
\alias{autoplot.PCA}
\alias{autoplot.pca}
\alias{autoplot.pcaRes}
\alias{autoplot.prcomp}
\alias{autoplot.rda}
\alias{autoplot_pca}
\title{}
\usage{
\method{autoplot}{prcomp}(object, mapping = aes(), type = c("observations",
  "variables"), data = NULL, PC = c(1, 2), scaling = "auto", ...)

\method{autoplot}{PCA}(object, mapping = aes(), type = c("observations",
  "variables"), data = NULL, PC = c(1, 2), scaling = "auto", ...)

\method{autoplot}{rda}(object, mapping = aes(), type = c("observations",
  "variables"), data = NULL, PC = c(1, 2), scaling = "auto", ...)

\method{autoplot}{pca}(object, mapping = aes(), type = c("observations",
  "variables"), data = NULL, PC = c(1, 2), scaling = "auto", ...)

\method{autoplot}{pcaRes}(object, mapping = aes(), type = c("observations",
  "variables"), data = NULL, PC = c(1, 2), scaling = "auto", ...)
}
\arguments{
\item{object}{an object resulting from a PCA}

\item{mapping}{a call to aes() specifying additional mappings between variables and plot aesthetics; by default, positions in x and y are mapped to the scores on the principal components and colour is mapped to the nature of the data (active or supplementary) when relevant. See \code{\link{fortify_pca}} for a list of the other mappable variables returned by the \code{fortify} methods}

\item{type}{the elements to consider: either "observations", "individuals", "sites", "lines" (which are all synonyms) or "variables", "descriptors", "species", "columns" (which are, again, synonyms)}

\item{data}{the original dataset, to be concatenated to the output when extracting observations. Whhen \code{NULL}, the default, data will be extracted from the PCA object when it contains it (i.e. not for \code{\link{prcomp}})}

\item{PC}{the principal components to extract; two is usual for plotting}

\item{scaling}{how to scale scores. Can be "none", "obs", "var", "both" (or synonyms of "obs" and "var", as for the argument \code{type}). Can also be a integer from 0 to 3, which designs a scaling as in the order above (0="none", 1="obs", etc.). Scaling observations makes the distances in the plot of observations approximations of the euclidan distance in the original space. Scaling variables relates the angle between variables to the correlations between them (right angle = no correlation). Scaling of observation or variables is by the square root of the eigenvalues. Scaling both is by the fourth root of the eigenvalues.
TODO add a note about auto scaling}

\item{...}{passed to the various geoms; can be used to \emph{set} further aesthetics}
}
\value{
A ggplot2 object defining the plot
}
\description{

}
\examples{
# PCA with stats::prcomp
pca <- prcomp(USArrests, scale = TRUE)
autoplot(pca)

# add further aesthetic mappings
names(fortify(pca, type = "obs"))
autoplot(pca, type = "obs", mapping=aes(alpha=.cos2))
autoplot(pca, type = "obs", mapping=aes(alpha=.cos2, size=.contrib))
# including from the original data
autoplot(pca, type = "obs", mapping=aes(alpha=.cos2, size=Murder), data=USArrests)

# aesthetics can also be set
autoplot(pca, type = "obs", mapping=aes(alpha=.cos2), colour="red", size=3, shape=15)

\dontrun{
# PCA with FactoMineR::PCA
library("FactoMineR")
# use supplementary observations and variables
pca <- PCA(USArrests, scale = TRUE, graph=FALSE, ind.sup = 2, quanti.sup = 4)

# colour is mapped by default
autoplot(pca)

# but the mapping can be overridden by mapping another variable
autoplot(pca, type = "obs", mapping = aes(colour=.contrib))
# or setting the corresponding aesthetic
autoplot(pca, type = "obs", colour="black")

# additional mappings can be specified
autoplot(pca, type = "obs", mapping = aes(colour=.contrib, alpha=.cos2, shape=.kind))
# in particular, with FactoMineR, the data is present by default and can be mapped
names(fortify(pca, type = "obs"))
autoplot(pca, aes(alpha=.cos2, size=Murder), "obs")

# PCA with pcaMethods::pca, from bioconductor
library("pcaMethods")
# settings equivalent `prcomp`
pca <- pca(USArrests, method="svd", scale="uv", completeObs=TRUE, nPcs=4)
autoplot(pca)
autoplot(pca, type="obs", mapping=aes(alpha=.cos2))

# other computation method specific to pcaMethods
pca <- pca(USArrests, method="ppca", scale="uv", completeObs=TRUE, nPcs=4)
autoplot(pca)

# PCA with vegan::rda
library("vegan")
pca <- rda(USArrests, scale = TRUE)
autoplot(pca)

}
}
\author{
Jean-Olivier Irisson \email{irisson@normalesup.org}
}
\seealso{
\code{\link{fortify_pca}} for the function which actually extracts the data from the PCA object.

\code{\link{prcomp}} in package \code{stats}, \code{\link[FactoMineR:PCA]{PCA}} in package \code{factoMineR}, \code{\link[ade4:dudi.pca]{dudi.pca}} in package \code{ade4}, \code{\link[vegan:rda]{rda}} in package \code{vegan},   \code{\link[pcaMethods:pca]{pca}} in package \code{pcaMethods} from bioconductor.
}

