% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordination_eigenvalues.R
\name{eigenvalues}
\alias{eigenvalues}
\alias{eigenvalues.CA}
\alias{eigenvalues.PCA}
\alias{eigenvalues.ca}
\alias{eigenvalues.correspondence}
\alias{eigenvalues.pca}
\alias{eigenvalues.pcaRes}
\alias{eigenvalues.prcomp}
\alias{eigenvalues.rda}
\title{Eigenvalues in an ordination analysis}
\usage{
eigenvalues(x)

\method{eigenvalues}{prcomp}(x)

\method{eigenvalues}{PCA}(x)

\method{eigenvalues}{rda}(x)

\method{eigenvalues}{pca}(x)

\method{eigenvalues}{pcaRes}(x)

\method{eigenvalues}{CA}(x)

\method{eigenvalues}{correspondence}(x)

\method{eigenvalues}{ca}(x)
}
\arguments{
\item{x}{an object returned by an ordination function.}
}
\value{
A numeric vector containing the eigenvalues.
}
\description{
Extract eigenvalues from an ordination analysis (Principal Components Analysis, Correspondence Analysis, etc.) object in a standard way.
}
\examples{
# Principal Component Analysis
pca <- prcomp(USArrests, scale=TRUE)
eigenvalues(pca)
if (require("FactoMineR")) {
  eigenvalues(PCA(USArrests, graph=F))
}
if (require("vegan")) {
  eigenvalues(rda(USArrests, scale=TRUE))
}
if (require("ade4")) {
  eigenvalues(dudi.pca(USArrests, scannf=FALSE, nf=4))
}
if (require("pcaMethods")) {
  eigenvalues(pca(USArrests, scale="uv", nPcs=4))
}

# Correspondence analysis
clr <- HairEyeColor[,,1]
if (require("FactoMineR")) {
  eigenvalues(CA(clr, graph=F))
}
if (require("MASS")) {
  eigenvalues(corresp(clr, nf=3))
}
if (require("ca")) {
  eigenvalues(ca(clr))
}

}
\seealso{
Functions to perform PCA: \code{\link[stats]{prcomp}} in package \code{stats}, \code{\link[FactoMineR]{PCA}} in package \code{factoMineR}, \code{\link[vegan]{rda}} in package \code{vegan}, \code{\link[ade4]{dudi.pca}} in package \code{ade4}, \code{\link[pcaMethods]{pca}} in package \code{pcaMethods} (on bioconductor).

Functions to perform CA: \code{\link[FactoMineR]{CA}} in package \code{factoMineR}, \code{\link[MASS]{corresp}} in package \code{MASS}, \code{\link[ca]{ca}} in package \code{ca}.

Other CA-related functions: \code{\link{scores}}

Other PCA-related functions: \code{\link{autoplot_pca}},
  \code{\link{pca_tidiers}}, \code{\link{scores}}
}

