% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_tidiers.R
\name{pca_tidiers}
\alias{augment.PCA}
\alias{augment.pca}
\alias{augment.pcaRes}
\alias{augment.prcomp}
\alias{augment.rda}
\alias{pca_tidiers}
\alias{tidy.PCA}
\alias{tidy.pca}
\alias{tidy.pcaRes}
\alias{tidy.prcomp}
\alias{tidy.rda}
\title{Tidying methods for a Principal Component Analysis}
\usage{
tidy(x, ...)

augment(x, data=NULL, dimensions=c(1,2), which="row", scaling=which, ...)
}
\arguments{
\item{x}{an object returned by a function performing Principal Component Analysis.}

\item{...}{ignored.}

\item{data}{the original dataset, to be concatenated with the output when extracting row scores. When \code{NULL} (the default) data will be extracted from the PCA object when it contains it (i.e. for all functions but \code{\link[stats]{prcomp}}).}

\item{dimensions}{vector giving the indexes of the principal components to extract. Typically two are extracted to create a plot.}

\item{which}{the type of coordinates in the new space to extract: either "rows", "lines", "observations", "objects", "individuals", "sites" (which are all treated as synonyms) or "columns", "variables", "descriptors", "species" (which are, again, synonyms). All can be abbreviated. By default, coordinates of rows are returned. Row coordinates are commonly called 'scores' and column coordinates usually called 'loadings'.}

\item{scaling}{scaling for the scores. Can be
\describe{
  \item{"none" (or 0)}{for raw scores,}
  \item{"rows" (or 1, or a synonym of "rows")}{to scale row scores by the eigenvalues,}
  \item{"columns" (or 2, or a synonym of "columns")}{to scale column scores by the eigenvalues,}
  \item{"both" (or 3)}{to scale both row and column scores.}
}
By default, scaling is adapted to the type of scores extracted (scaling 1 for row scores, scaling 2 for column scores, and scaling 3 when scores are extracted for a biplot).}
}
\value{
For \code{tidy}, a data.frame containing the variance (i.e. eigenvalue), the proportion of variance, and the cumulative proportion of variance associated to each principal component.

For \code{augment}, a data.frame containing the original data (when \code{which="rows"} and \code{data} is supplied or can be extracted from the object) and the additional columns:
\describe{
  \item{.rownames:}{the identifier of the row or column, extracted from the row or column names in the original data.}
  \item{.PC#:}{the coordinates of data objects on the extracted principal components.}
  \item{.cos2:}{the squared cosine, summed over extracted PCs, which quantifies the quality of the representation of each data point in the space of the extracted PCs. NB: \code{cos2} can only be computed when all possible principal components are extracted in the PCA objects; when it is not the case, \code{cos2} is \code{NA}. In several packages, the number of principal components to keep is an argument of the PCA function (and the default is not "all").}
  \item{.contrib:}{the contribution of each object to the selected PCs. NB: same comment as for \code{cos2} regarding the number of PCs kept in the PCA object.}
  \item{.type:}{the nature of the data extracted : \code{row} or \code{col}.}
}
}
\description{
Extract diagnostics, coordinates on the principal components (i.e. rows scores and columns loadings) and some fit statistics from a Principal Component Analysis.
}
\details{
Scaling of scores follows the conventions of package \code{vegan}. In summary, except for scaling 0, scores are all multiplied by a constant:
  \deqn{c = \sqrt[4]{(n-1) \times \sum{eig}}}{c = sqrt(sqrt((n-1) * sum(eig)))}
where \eqn{n} is the number of active rows in the ordination and \eqn{eig} are the eigenvalues.
In addition, for scaling 1 (resp. 2), row scores (resp. column scores) are multiplied by:
  \deqn{\sqrt{\frac{eig}{\sum{eig}}}}{sqrt(eig/sum(eig))}
For scaling 3, both row and column scores are multiplied by:
  \deqn{\sqrt[4]{\frac{eig}{\sum{eig}}}}{sqrt(sqrt(eig/sum(eig)))}
For details and justification, see \code{vignette("decision-vegan")}.
}
\examples{
pca <- prcomp(USArrests, scale = TRUE)

tidy(pca)

head(augment(pca, which="row"))
head(augment(pca, which="col"))
# or use your preferred synonym, possibly abbreviated
head(augment(pca, which="obs"))
head(augment(pca, which="var"))
head(augment(pca, which="descriptors"))

# data is not contained in the `prcomp` object but can be provided
head(augment(pca, data=USArrests, which="row"))
# select different principal components
augment(pca, which="col", dim=c(2,3))

if (require("FactoMineR")) {
  pca <- FactoMineR::PCA(USArrests, graph=FALSE, ncp=4)
  head(augment(pca, which="individuals"))
  head(augment(pca, which="variables"))
}

if (require("vegan")) {
  pca <- vegan::rda(USArrests, scale=TRUE)
  # can use vegan's naming convention
  head(augment(pca, which="sites"))
  head(augment(pca, which="species"))
}

if (require("ade4")) {
  pca <- ade4::dudi.pca(USArrests, scannf=FALSE)
  head(augment(pca))
  head(augment(pca, which="variables"))
}

if (require("pcaMethods")) {
  pca <- pcaMethods::pca(USArrests, scale="uv")
  head(augment(pca))
  augment(pca, which="var")
}

}
\seealso{
Functions to perform PCA: \code{\link[stats]{prcomp}} in package \code{stats}, \code{\link[FactoMineR]{PCA}} in package \code{factoMineR}, \code{\link[vegan]{rda}} in package \code{vegan}, \code{\link[ade4]{dudi.pca}} in package \code{ade4}, \code{\link[pcaMethods]{pca}} in package \code{pcaMethods} (on bioconductor).

Other PCA.related.functions: \code{\link{autoplot_pca}},
  \code{\link{ca_tidiers}}
}

